+++ CHANGELOG.md
@@ -8,12 +8,15 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - peer lists accepts a `DefaultChooseTimeout` configuration for applying to
   `context`s without deadlines.
+- gRPC-go version is added to debug pages.
 ### Fixed
 - yarpcerrors: `fmt` verbs are ignored when no args are passed to error
   constructors.
 - Fix gRPC streaming when used with the direct peer chooser.
 - Streaming calls do not require contexts with deadlines. Users should use
   cancelable contexts for long-lived streams instead of timeouts.
+## Removed
+- The `x/yarpcmeta` package is completely removed.
 
 ## [1.45.0] - 2020-04-21
 ### Added